refactor: extract container config and environment helpers from instance.rs
Splits the 2246-line instance.rs into focused modules:

- container_config.rs (1103 lines): sandbox directory sync, agent config
  mounting, credential extraction, volume path computation, and container
  config building with all associated tests.
- environment.rs (179 lines): shell escaping, env var collection, docker
  exec env flag building with all associated tests.
- instance.rs (986 lines): session lifecycle, tmux integration, and
  Instance struct/methods.

Also consolidates the three apply_*_tmux_options methods into a shared
apply_session_tmux_options helper with thin wrappers, and removes the
compute_volume_paths wrapper method on Instance in favor of calling the
free function directly.
