+++ src/session/mod.rs
@@ -3,6 +3,8 @@
 pub mod builder;
 pub mod civilizations;
 pub mod config;
+mod container_config;
+mod environment;
 mod groups;
 mod instance;
 pub mod profile_config;
